requirements.txt
alembic==1.1.0
astroid==2.2.5
autopep8==1.4.4
Babel==2.7.0
bleach>=3.3.0
Click==7.0
colorama==0.4.1
Flask==1.1.1
Flask-Babel==0.12.2
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
Flask-Table==0.5.0
Flask-WTF==0.14.2
gunicorn==19.9.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.4.1
login==0.0.6
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2-binary==2.8.3
pycodestyle==2.5.0
pylint==2.3.1
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.6
typed-ast==1.5.0
webencodings==0.5.1
Werkzeug==0.15.5
wrapt==1.11.2
WTForms==2.2.1
